pyproject.toml
[tool.poetry]
name = "sdkcpc"
version = "0.1.0"
description = ""
authors = ["Destroyer <destroyer.dcf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
rich = "12.0.0"
DateTime = "^4.4"
PyYAML = "^6.0"
Jinja2 = "^3.1.2"
ipaddress = "^1.0.23"
tabulate = "^0.8.9"
inquirer = "^2.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
